+++ src/de/omegazirkel/risingworld/tools/OZLogger.java
@@ -139,12 +139,13 @@ public class OZLogger {
         isInShutdownMode = false;
     }
 
-    public void setLevel(String levelName){
-        logger().setLevel(Level.toLevel(levelName));
+    public OZLogger setLevel(String levelName){
+        return setLevel(Level.toLevel(levelName));
     }
 
-    public void setLevel(Level level) {
+    public OZLogger setLevel(Level level) {
         logger().setLevel(level);
+        return this;
     }
 
     public void debug(String message) {